requirements.txt
Django==1.7.4
coverage==3.7.1
django-nose==1.3
django-rest==0.0.1
djangorestframework==3.0.4
mock==1.0.1
nose==1.3.4
pep8==1.5.7
wsgiref==0.1.2
